### 2022-09-24 019b092
Added JOSS badge to readme
