### 2022-10-03 f4dc692
added new reaction calculator
