### 2022-10-22 a6ff3e3
sanitised solution init
