+++ docs/changelog/20221022_bobmyhill.rst
@@ -1,0 +1,7 @@
+* BurnMan Solution and ElasticSolution objects are now instantiated
+  with a SolutionModel object as a solution_model parameter.
+  The use of the solution_type parameter has been removed
+  completely, along with all of the optional parameters that were
+  originally passed as parameters to SolutionModel.
+
+  *Bob Myhill, 2022/10/22*

[…]
