+++ burnman/data/input_raw_endmember_datasets/HPdata_to_burnman.py
@@ -14,92 +14,133 @@ import sys
 import os.path
 import pprint
 from collections import OrderedDict
+
 # hack to allow scripts to be placed in subdirectories next to burnman:
-if not os.path.exists('burnman') and os.path.exists('../../../burnman'):
-    sys.path.insert(1, os.path.abspath('../../..'))
+if not os.path.exists("burnman") and os.path.exists("../../../burnman"):
+    sys.path.insert(1, os.path.abspath("../../.."))
 
 from burnman.tools.chemistry import dictionarize_formula, formula_mass
 
 
-if not os.path.isfile('tc-ds62.txt'):
-    print('This code requires the data file tc-ds62.txt.')
-    print('This file is bundled with the software package THERMOCALC, '
-          'which can be found here:')
-    print('http://www.metamorph.geo.uni-mainz.de/thermocalc/'
-          'dataset6/index.html')
-    print('')
-    print('Please download the file and place it in this directory.')
+if not os.path.isfile("tc-ds62.txt"):
+    print("This code requires the data file tc-ds62.txt.")
+    print(
+        "This file is bundled with the software package THERMOCALC, "
+        "which can be found here:"
+    )
+    print("http://www.metamorph.geo.uni-mainz.de/thermocalc/" "dataset6/index.html")
+    print("")
+    print("Please download the file and place it in this directory.")
     exit()
 
 # Components
-components = ['Si', 'Ti', 'Al', 'Fe', 'Mg', 'Mn', 'Ca', 'Na',
-              'K', 'O', 'H', 'C', 'Cl', 'e-', 'Ni', 'Zr', 'S', 'Cu', 'Cr']
+components = [
+    "Si",
+    "Ti",
+    "Al",
+    "Fe",
+    "Mg",
+    "Mn",
+    "Ca",
+    "Na",
+    "K",
+    "O",
+    "H",
+    "C",
+    "Cl",
+    "e-",
+    "Ni",
+    "Zr",
+    "S",
+    "Cu",
+    "Cr",
+]
 
 
 class Endmember:
-
-    def __init__(self, name, atoms, formula, sites, comp,
-                 H, S, V, Cp, a, k, flag, od):
+    def __init__(self, name, atoms, formula, sites, comp, H, S, V, Cp, a, k, flag, od):
         if flag != -1 and flag != -2 and k[0] > 0:
             formula = dictionarize_formula(formula)
-            self.params = OrderedDict([('name', name),
-                                       ('formula', formula),
-                                       ('equation_of_state', 'hp_tmt'),
-                                       ('H_0', round(H * 1e3, 10)),
-                                       ('S_0', round(S * 1e3, 10)),
-                                       ('V_0', round(V * 1e-5, 15)),
-                                       ('Cp', [round(Cp[0] * 1e3, 10),
-                                               round(Cp[1] * 1e3, 10),
-                                               round(Cp[2] * 1e3, 10),
-                                               round(Cp[3] * 1e3, 10)]),
-                                       ('a_0', a),
-                                       ('K_0', round(k[0] * 1e8, 10)),
-                                       ('Kprime_0', k[1]),
-                                       ('Kdprime_0', round(k[2] * 1e-8, 15)),
-                                       ('n', sum(formula.values())),
-                                       ('molar_mass',
-                                        round(formula_mass(formula), 10))])
+            self.params = OrderedDict(
+                [
+                    ("name", name),
+                    ("formula", formula),
+                    ("equation_of_state", "hp_tmt"),
+                    ("H_0", round(H * 1e3, 10)),
+                    ("S_0", round(S * 1e3, 10)),
+                    ("V_0", round(V * 1e-5, 15)),
+                    (
+                        "Cp",
+                        [
+                            round(Cp[0] * 1e3, 10),
[…]
