+++ burnman/eos/birch_murnaghan.py
@@ -1,4 +1,5 @@
 from __future__ import absolute_import
+
 # This file is part of BurnMan - a thermoelastic and thermodynamic toolkit for
 # the Earth and Planetary Sciences
 # Copyright (C) 2012 - 2017 by the BurnMan team, released under the GNU
@@ -19,11 +20,18 @@ def bulk_modulus(volume, params):
     modulus.  Pressure must be in :math:`[Pa]`.
     """
 
-    x = params['V_0'] / volume
-    f = 0.5 * (pow(x, 2. / 3.) - 1.0)
+    x = params["V_0"] / volume
+    f = 0.5 * (pow(x, 2.0 / 3.0) - 1.0)
 
-    K = pow(1. + 2. * f, 5. / 2.) * (params['K_0'] + (3. * params['K_0'] * params['Kprime_0'] -
-                                                      5 * params['K_0']) * f + 27. / 2. * (params['K_0'] * params['Kprime_0'] - 4. * params['K_0']) * f * f)
+    K = pow(1.0 + 2.0 * f, 5.0 / 2.0) * (
+        params["K_0"]
+        + (3.0 * params["K_0"] * params["Kprime_0"] - 5 * params["K_0"]) * f
+        + 27.0
+        / 2.0
+        * (params["K_0"] * params["Kprime_0"] - 4.0 * params["K_0"])
+        * f
+        * f
+    )
     return K
 
 
@@ -34,8 +42,14 @@ def birch_murnaghan(x, params):
     modulus (params['K_0'])
     """
 
-    return 3. * params['K_0'] / 2. * (pow(x, 7. / 3.) - pow(x, 5. / 3.)) \
-        * (1. - .75 * (4. - params['Kprime_0']) * (pow(x, 2. / 3.) - 1.)) + params['P_0']
+    return (
+        3.0
+        * params["K_0"]
+        / 2.0
+        * (pow(x, 7.0 / 3.0) - pow(x, 5.0 / 3.0))
+        * (1.0 - 0.75 * (4.0 - params["Kprime_0"]) * (pow(x, 2.0 / 3.0) - 1.0))
+        + params["P_0"]
+    )
 
 
 def volume(pressure, params):
@@ -44,12 +58,13 @@ def volume(pressure, params):
     pressure :math:`[Pa]`. Returns molar volume in :math:`[m^3]`
     """
 
-    func = lambda x: birch_murnaghan(params['V_0'] / x, params) - pressure
+    func = lambda x: birch_murnaghan(params["V_0"] / x, params) - pressure
     try:
-        sol = bracket(func, params['V_0'], 1.e-2 * params['V_0'])
+        sol = bracket(func, params["V_0"], 1.0e-2 * params["V_0"])
     except:
         raise ValueError(
-            'Cannot find a volume, perhaps you are outside of the range of validity for the equation of state?')
+            "Cannot find a volume, perhaps you are outside of the range of validity for the equation of state?"
+        )
     return opt.brentq(func, sol[0], sol[1])
 
 
@@ -60,9 +75,17 @@ def shear_modulus_second_order(volume, params):
     params['G_0']).  This uses a second order finite strain expansion
     """
 
-    x = params['V_0'] / volume
-    G = params['G_0'] * pow(x, 5. / 3.) * (
-        1. - 0.5 * (pow(x, 2. / 3.) - 1.) * (5. - 3. * params['Gprime_0'] * params['K_0'] / params['G_0']))
+    x = params["V_0"] / volume
+    G = (
+        params["G_0"]
+        * pow(x, 5.0 / 3.0)
+        * (
+            1.0
+            - 0.5
+            * (pow(x, 2.0 / 3.0) - 1.0)
[…]

[…]
