+++ burnman/version.py
@@ -1,2 +1,2 @@
-version = "1.2.0pre"
+version = "1.2.0a0"
 short_version = "1.2.0"
