### 2022-10-22 f5e63a7
polynomial solution model
