+++ burnman/classes/solutionmodel.py
@@ -10,6 +10,10 @@ import numpy as np
 from ..utils.chemistry import process_solution_chemistry
 from .. import constants
 import warnings
+import sparse
+import string
+from copy import deepcopy
+from .material import material_property, cached_property
 
 try:
     ag = importlib.import_module("autograd")
@@ -334,6 +338,27 @@ class SolutionModel(object):
         """
         return np.zeros_like(np.array(molar_fractions))
 
+    def Cp_excess(self):
+        """
+        Returns the excess heat capacity of the solution model
+        at its current state
+        """
+        return 0.0
+
+    def alphaV_excess(self):
+        """
+        Returns the excess alpha*V of the solution model
+        at its current state
+        """
+        return 0.0
+
+    def VoverKT_excess(self):
+        """
+        Returns the excess V/K_T of the solution model
+        at its current state
+        """
+        return 0.0
+
 
 class MechanicalSolution(SolutionModel):
 
@@ -476,7 +501,6 @@ class IdealSolution(SolutionModel):
             self.endmember_noccupancies,
             logish(site_noccupancies) - logish(site_multiplicities),
         )
-
         normalisation_constants = (
             self.endmember_configurational_entropies / constants.gas_constant
         )
@@ -1021,6 +1045,456 @@ class FunctionSolution(IdealSolution):
             pressure, temperature, molar_fractions
         )
         return ideal_entropy_hessian + nonideal_entropy_hessian
+
+    def activity_coefficients(self, pressure, temperature, molar_fractions):
+        if temperature > 1.0e-10:
+            return np.exp(
+                self._non_ideal_excess_partial_gibbs(
+                    pressure, temperature, molar_fractions
+                )
+                / (constants.gas_constant * temperature)
+            )
+        else:
+            raise Exception("Activity coefficients not defined at 0 K.")
+
+    def activities(self, pressure, temperature, molar_fractions):
+        return IdealSolution.activities(
+            self, pressure, temperature, molar_fractions
+        ) * self.activity_coefficients(pressure, temperature, molar_fractions)
+
+
+class PolynomialSolution(IdealSolution):
+    """
+    Solution model implementing a general polynomial solution model.
+
+    Parameters
+    ----------
+    endmembers : list of lists
+        A list of all the independent endmembers in the solution.
+        The first item of each list gives the Mineral object corresponding
+        to the endmember. The second item gives the site-species formula.
+
+    ESV_interactions : list of lists
+        A list containing lists where the first three elements are
+        energy, entropy and volume interactions and the rest of the elements
+        are indices of the transformed endmembers to which those
+        interactions correspond.
+        For example, [2., 0., 0., 0, 1, 1] would correspond to an interaction
+        of 2*p'[0]*p'[1]*p'[1].
+
+    interaction_endmembers : list of burnman.Mineral objects
+        A list of minerals involved in the interaction terms.
+
+    endmember_coefficients_and_interactions : list of lists
+        A list containing lists where the first n elements are
+        coefficients for each of the interaction_endmembers and the
+        rest of the elements are indices of the transformed
+        endmembers
+        to which those interactions correspond.
[…]
