### 2022-10-22 d6489cd
new tests for polynomial solution model
