+++ tests/test_solidsolution.py
@@ -13,6 +13,7 @@ from burnman.classes.solutionmodel import SymmetricRegularSolution
 from burnman.classes.solutionmodel import SubregularSolution
 from burnman.classes.solutionmodel import AsymmetricRegularSolution
 from burnman.classes.solutionmodel import FunctionSolution
+from burnman.classes.solutionmodel import PolynomialSolution
 from burnman.utils.chemistry import dictionarize_formula, formula_mass
 from burnman.utils.chemistry import formula_to_string, sum_formulae
 from burnman.minerals import HGP_2018_ds633
@@ -67,9 +68,8 @@ class fayalite(Mineral):
 
 made_forsterite = CombinedMineral([forsterite(), forsterite()], [0.5, 0.5])
 
+
 # One-mineral solid solution
-
-
 class forsterite_ss(burnman.SolidSolution):
     def __init__(self, molar_fractions=None):
         self.name = "Dummy solid solution"
@@ -81,8 +81,6 @@ class forsterite_ss(burnman.SolidSolution):
 
 
 # Two-mineral solid solution
-
-
 class forsterite_forsterite_ss(burnman.SolidSolution):
     def __init__(self, molar_fractions=None):
         self.name = "Fo-Fo solid solution"
@@ -95,8 +93,6 @@ class forsterite_forsterite_ss(burnman.SolidSolution):
 
 
 # Ideal solid solution
-
-
 class olivine_ideal_ss(burnman.SolidSolution):
     def __init__(self, molar_fractions=None):
         self.name = "Fo-Fo solid solution"
@@ -108,8 +104,6 @@ class olivine_ideal_ss(burnman.SolidSolution):
 
 
 # Olivine solid solution
-
-
 class olivine_ss(burnman.SolidSolution):
     def __init__(self, molar_fractions=None):
         self.name = "Olivine"
@@ -134,8 +128,6 @@ class olivine_ss2(burnman.SolidSolution):
 
 
 # Orthopyroxene solid solution
-
-
 class orthopyroxene(burnman.SolidSolution):
     def __init__(self, molar_fractions=None):
         # Name
@@ -152,8 +144,6 @@ class orthopyroxene(burnman.SolidSolution):
 
 
 # Three-endmember, two site symmetric solid solution
-
-
 class two_site_ss(burnman.SolidSolution):
     def __init__(self, molar_fractions=None):
         self.name = "two_site_ss"
@@ -170,8 +160,6 @@ class two_site_ss(burnman.SolidSolution):
 
 
 # Three-endmember, two site asymmetric solid solution
-
-
 class two_site_ss_asymmetric(burnman.SolidSolution):
     def __init__(self, molar_fractions=None):
         self.name = "two_site_ss (asymmetric)"
@@ -189,11 +177,8 @@ class two_site_ss_asymmetric(burnman.SolidSolution):
 
 
 # Three-endmember, two site solid solution
-
-
 class two_site_ss_subregular(burnman.SolidSolution):
     def __init__(self, molar_fractions=None):
-        # Name
         self.name = "two_site_ss (subregular symmetric)"
         self.solution_model = SubregularSolution(
             endmembers=[
@@ -213,7 +198,6 @@ class two_site_ss_subregular(burnman.SolidSolution):
 # Three-endmember, two site solid solution
 class two_site_ss_subregular_asymmetric(burnman.SolidSolution):
     def __init__(self, molar_fractions=None):
-        # Name
         self.name = "two_site_ss (subregular symmetric)"
[…]
