### 2022-10-22 90fb8eb
add sparse to reqs
