+++ docs/pip_requirements.txt
@@ -8,6 +8,7 @@ nbsphinx
 cvxpy
 jupyter
 autograd
+sparse
 
 # ReadTheDocs cannot build pycddlib successfully
 # because it requires gmp.h, installable only via sudo apt-get install.
