+++ burnman/eos/einstein.py
@@ -7,6 +7,21 @@ from __future__ import absolute_import
 import numpy as np
 from .. import constants
 
+# Try to import the jit from numba.  If it is
+# not available, just go with the standard
+# python interpreter
+try:
+    import os
+
+    if "NUMBA_DISABLE_JIT" in os.environ and int(os.environ["NUMBA_DISABLE_JIT"]) == 1:
+        raise ImportError("NOOOO!")
+    from numba import jit
+except ImportError:
+
+    def jit(fn):
+        return fn
+
+
 """
 Functions for the Einstein model of a solid.
 """
@@ -14,6 +29,7 @@ Functions for the Einstein model of a solid.
 eps = np.finfo(float).eps
 
 
+@jit
 def thermal_energy(T, einstein_T, n):
     """
     calculate the thermal energy of a substance.  Takes the temperature,
@@ -21,15 +37,13 @@ def thermal_energy(T, einstein_T, n):
     Returns thermal energy in J/mol
     """
     if T <= eps:
-        # zero point energy
-        return 3.0 * n * constants.gas_constant * einstein_T * 0.5
+        return 0.0
     x = einstein_T / T
-    E_th = (
-        3.0 * n * constants.gas_constant * einstein_T * (0.5 + 1.0 / (np.exp(x) - 1.0))
-    )  # include the zero point energy
+    E_th = 3.0 * n * constants.gas_constant * einstein_T * (1.0 / (np.exp(x) - 1.0))
     return E_th
 
 
+@jit
 def molar_heat_capacity_v(T, einstein_T, n):
     """
     Heat capacity at constant volume.  In J/K/mol
@@ -44,3 +58,57 @@ def molar_heat_capacity_v(T, einstein_T, n):
         * (x * x * np.exp(x) / np.power(np.exp(x) - 1.0, 2.0))
     )
     return C_v
+
+
+@jit
+def helmholtz_free_energy(T, einstein_T, n):
+    """
+    Helmholtz free energy of lattice vibrations in the Einstein model [J].
+    It is important to note that this does NOT include the zero
+    point energy for the lattice.  As long as you are
+    calculating relative differences in F, this should cancel anyway.
+    """
+    E = thermal_energy(T, einstein_T, n)
+    S = entropy(T, einstein_T, n)
+    return E - T * S
+
+
+@jit
+def entropy(T, einstein_T, n):
+    """
+    Entropy due to lattice vibrations in the Einstein model [J/K]
+    """
+    if T <= eps:
+        return 0.0
+    x = einstein_T / T
+    S = (
+        3.0
+        * n
+        * constants.gas_constant
+        * (-x * np.exp(-x) / (np.exp(-x) - 1.0) - np.log(1.0 - np.exp(-x)))
+    )
+    return S
+
+
+@jit
+def dmolar_heat_capacity_v_dT(T, einstein_T, n):
+    """
+    First temperature derivative of the heat capacity at constant volume
+    according to the Einstein model [J/K^2/mol].
+    """
+    if T <= eps:
+        return 0.0
+
+    x = einstein_T / T
+    dCvdT = (
+        3.0
+        * n
+        * constants.gas_constant
+        * x
+        * x
+        * np.exp(x)
+        * ((x - 2.0) * np.exp(x) + (x + 2.0))
+        / (T * np.power(np.exp(x) - 1.0, 3.0))
+    )
+
+    return dCvdT
