+++ docs/changelog/20221023_bobmyhill.rst
@@ -1,0 +1,12 @@
+* Four new property modifier formulations are provided, 
+  which can be specified with the names "debye", "debye_delta", 
+  "einstein" and "einstein_delta". These 
+  are based on the Debye and Einstein models of thermal energy.
+  The heat capacity ("debye", "einstein")
+  or entropy ("debye_delta", "einstein_delta")
+  are based on the heat capacity of the respective thermal model,
+  and so reach a maximum at high temperature. The
+  excess energy, entropy and heat capacity are all zero
+  at 0 K. The excess volume is always zero.
+
+  *Bob Myhill, 2022/10/23*
