### 2022-10-24 935899a
used analytical derivative in anisotropic eos
