### 2022-10-26 d5181dc
fixed root finding
