### 2022-10-27 1fd499d
created anisotropy utils
