+++ burnman/classes/anisotropicmineral.py
@@ -7,54 +7,15 @@ from scipy.linalg import expm, logm
 from numpy.linalg import cond
 from .mineral import Mineral
 from .material import Material, material_property
-from .anisotropy import AnisotropicMaterial, voigt_compliance_factors
+from .anisotropy import AnisotropicMaterial
 from ..utils.misc import copy_documentation
 from ..utils.unitcell import cell_parameters_to_vectors
 from ..utils.unitcell import cell_vectors_to_parameters
-
-
-def contract_stresses(stresses):
-    """
-    Takes a stress tensor in standard (3x3) form
-    and returns the Voigt form (6). No factors
-    are required to maintain the relationship
-    with the corresponding stiffness and strain tensors
-    (see contract_strains, which does require
-    multiplicative factors).
-    """
-    return stresses[[0, 1, 2, 1, 0, 0], [0, 1, 2, 2, 2, 1]]
-
-
-def expand_stresses(stresses):
-    """
-    Takes a stress tensor in Voigt form (6)
-    and returns the standard form (3x3). No factors
-    are required to maintain the relationship
-    with the corresponding stiffness and strain tensors
-    (see contract_strains, which does require
-    multiplicative factors).
-    """
-    return np.array(
-        [
-            [stresses[0], stresses[5], stresses[4]],
-            [stresses[5], stresses[1], stresses[3]],
-            [stresses[4], stresses[3], stresses[2]],
-        ]
-    )
-
-
-def contract_strains(strains):
-    """
-    Takes a stress tensor in standard (3x3) form
-    and returns the Voigt form (6). Note the factors
-    which are required to maintain the relationship
-    with the corresponding stiffness and strain tensors.
-    """
-    # next line creates a copy, not just a view.
-    eps = strains[[0, 1, 2, 1, 0, 0], [0, 1, 2, 2, 2, 1]]
-    # only overwrites values in eps, not the input array
-    eps[3:] *= 2.0
-    return eps
+from ..utils.anisotropy import (
+    voigt_notation_to_compliance_tensor,
+    voigt_notation_to_stiffness_tensor,
+    contract_compliances,
+)
 
 
 class AnisotropicMineral(Mineral, AnisotropicMaterial):
@@ -146,9 +107,9 @@ class AnisotropicMineral(Mineral, AnisotropicMaterial):
                 f"by {factor}."
             )
 
-        self.Psi_0_Voigt = self._contract_compliances(
-            np.einsum("ij, kl", logm(self.cell_vectors_0), np.eye(3) / 3.0)
-        )
+        # Note, Psi_0 may be asymmetric, in which case the Voigt contraction
+        # cannot be applied
+        self.Psi_0 = np.einsum("ij, kl", logm(self.cell_vectors_0), np.eye(3) / 3.0)
 
         self.isotropic_mineral = isotropic_mineral
         if "name" in isotropic_mineral.params:
@@ -206,8 +167,8 @@ class AnisotropicMineral(Mineral, AnisotropicMaterial):
         self._f = f
 
         out = self.psi_function(f, self.Pth, self.anisotropic_params)
-        self.Psi_Voigt, self.dPsidf_Voigt, self.dPsidPth_Voigt = out
-        self.Psi_Voigt += self.Psi_0_Voigt
+        Psi_Voigt, self.dPsidf_Voigt, self.dPsidPth_Voigt = out
+        self.Psi = voigt_notation_to_compliance_tensor(Psi_Voigt) + self.Psi_0
 
         # Convert to (f, T) variables
         self.dPsidP_Voigt = -self.isothermal_compressibility_reuss * (
@@ -217,34 +178,6 @@ class AnisotropicMineral(Mineral, AnisotropicMaterial):
             self.dPsidf_Voigt
[…]
