+++ AUTHORS.md
@@ -1,0 +1,23 @@
+Project Authors
+===============
+
+BurnMan is developed by a collaborative and inclusive community.
+
+### Project technical lead
+
+* Bob Myhill (University of Bristol, UK)
+
+### Developers
+
+* Sanne Cottaar (University of Cambridge, UK)
+* Timo Heister (Clemson University, USA)
+* Ian Rose (Independent Researcher, USA)
+* Cayman Unterborn (Southwest Research Institute, USA)
+
+### Contributors
+
+A complete and growing list of the many authors that have contributed
+over the years can be found at [here](https://github.com/geodynamics/burnman/graphs/contributors).
+
+We are grateful for the contributions of these people, and
+their participation in our community!
