### 2023-04-19 #4797 @jedbrown
@anowacki @GeoscienceSoftware :wave:  Would you be available to review this submission to JOSS?
