### 2023-04-20 #5389 @editorialbot
[REVIEW]: BurnMan -- a Python toolkit for planetary geophysics, geochemistry and thermodynamics
**Submitting author:** @bobmyhill (<a href="http://orcid.org/0000-0001-9489-5236">Robert Myhill</a>)
**Repository:** https://github.com/geodynamics/burnman
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.2.0
