### 2023-04-24 #4797 @anowacki
Hi @jedbrown.  It's worth noting that I am currently working with @bobmyhill on a project, and so it's possible that may be viewed as a conflict of interest.

If you are happy to go ahead with me as a reviewer, then I would be willing to review this.  However, please also be aware that I won't be able to make much progress on this for a few weeks.  If that is not soon enough, then please do suggest others.

Finally, I am not an expert in Python and so whilst I can look at the API and documentation side of things, I am not that well placed to look at the tests or suggest improvements to general reproducibility.
