### 2023-05-25 geodynamics/burnman#526 @kaylai
Manuscript: Improve clarity on mineral phase proportion calculations
I was confused by Figure 5 of the manuscript showing mineral phase proportions computed from experimental phase equilibria data from Bertka & Fei (1997), and I think more clarity here will help.

**Issue:**
I originally thought that the phase proportions were calculated by BurnMan in a way similar to MELTS, where a minimization is performed to equilibrate some bulk composition at a given P-T-S-V (some combination of these), and then the residuals were errors based on direct comparison to Bertka & Fei's data. But, now I realize that you are performing a least squares calculation on bulk + individual phase data all reported by Berka & Fei (something we commonly do for these experiments, so this is useful!).

**Suggestion:**
I would clarify that this Figure is generated using the `fit_phase_proportions_to_bulk_composition()` method and write in the text that this is what is being done here, and that no thermodynamic calculation is being performed by BurnMan. Can you also explain what the residuals mean here (how they are calculated) either in the manuscript text or in the method's doc string?

**Further thought:**
I see from the doc string that there is a constraint that a phase cannot be negative. Can an individual component (e.g., oxide) go negative? This can be useful for determining if an experiment has lost some component, for example the loss of Fe to a platinum capsule. Often, when performing this calculation more manually, we add some pure Fe phase if we want to check for Fe loss and look to see if the least squares fit returns a negative value. This is mostly food for thought for you on how this tool can be extra useful for experimentalists.

In reference to: openjournals/joss-reviews#5389
