### 2023-05-25 geodynamics/burnman#527 @kaylai
Manuscript: Reference jupyter notebooks in tutorial directly in paper
**Issue:**
The jupyter notebooks in the tutorial folder give very complete and clear descriptions of points that are summarized in the manuscript itself. I did not realize there was much more information on the exact things discussed in the manuscript in those tutorial files until I went looking for them.

**Suggestion:**
I realize it will make the paper very long (and redundant with the tutorial) to completely port this text over to the manuscript. I think it would be very beneficial to directly reference specific jupyter notebooks in the manuscript text so that readers know precisely where to go to get more information and run the calculations themselves. I suggest doing this for each scenario rather than one sentence toward the end. This could be done in text or in figure captions as appropriate.

In reference to: https://github.com/openjournals/joss-reviews/issues/5389
