### 2023-05-25 geodynamics/burnman#528 @kaylai
Clarification on optional dependencies
cvxpy, pycddlib, and autograd are listed in the ReadMe as optional dependencies.

**Issues/questions:**
- What functionality will I lose if I do not install these?
- I was able to easily pip install cvxpy and autograd. Is there a reason you don't include these as standard dependencies that automatically get installed for me when I pip install BurnMan?
- ~~I got a legacy-install-failure error when I tried to pip install pycddlib. Is there another suggested way to install this?~~ *Fixed, see comment below.*

**Suggestions:**
Adding clarity to the ReadMe, specifically what functionality I will lose if I do not install these.

In reference to: https://github.com/openjournals/joss-reviews/issues/5389
