### 2023-05-25 geodynamics/burnman#528 @kaylai
Quick update, I worked harder to install pycddlib (I actually read the error message and then googled that...) I am doing this on my work computer, which has frequent updates. I assume one of them knocked out my previous install of XCode command line tools. Following the guidance [here](https://apple.stackexchange.com/questions/254380/why-am-i-getting-an-invalid-active-developer-path-when-attempting-to-use-git-a), I simply did:

`xcode-select --install`

and then tried `pip install pycddlib` again, and it worked. Given your docs also say to make sure XCode is installed, that one is on me!
