### 2023-05-25 geodynamics/burnman#529 @kaylai
Docstrings: Argument names are smooshed together with argument types
It seems in most (all?) of the autogenerated documentation from docstrings, the argument types are smooshed (this is a technical term) together with the argument type, so instead of

`myarg float`
or
```
myarg
    float
```

you get
`myargfloat`

which did confuse me quite a bit at fist in some cases!

In reference to: https://github.com/openjournals/joss-reviews/issues/5389
