### 2023-05-25 geodynamics/burnman#531 @kaylai
Manuscript: Text suggestions
Some suggestions for changes to the manuscript.

- Since BurnMan has been published in previous iterations, it would be useful to add even just a sentence explaining what the major update in functionality is to this version compared to previously published ones. Particularly for users wanting to cite BurnMan (i.e., if they have used BurnMan before they would have likely cited the 2014 G3 paper and maybe the newer zenodo). What is the difference between the version in the 2014 paper and this one?
- Second paragraph bulleted list: Use of "many" and "several" is vague. Can you state exactly how many?
- Figure 4 (Profile through planet Zog): I suggest defining PREM and being more specific in the caption, for example: "Depth dependent changes to density, gravity, pressure, and temperature are shown for Planet Zog (solid blue line) and compared with the Preliminary Reference Earth Model (PREM; dotted orange line) for density, pressure, and gravity. The geotherm of Planet Zog is compared to several from the literature." I would also use a different color for S1977 since a dotted orange line is used for PREM in the other plots.
- The Figure 5 caption needs to be clarified related to https://github.com/geodynamics/burnman/issues/526
- Should MELTS be included in the list of Gibbs minimization codes?

In reference to: https://github.com/openjournals/joss-reviews/issues/5389
