### 2023-05-30 geodynamics/burnman#528 @simonwmatthews
I agree that this would be a good thing to clarify.

This is not an issue with burnman, but it might be something that is useful to flag up... I had some issues installing cvxpy on an M1 chip mac, which appears to be a known issue with cvxpy. This problem also meant the test suite (test.sh) would not run properly, though the tutorials do (or at least appear to) run correctly.
