### 2023-06-06 geodynamics/burnman#533 @simonwmatthews
tutorial_01: missing example of ideal melt model
Halfway through tutorial 1 is this paragraph:

"As of BurnMan 1.1, ideality does not require that the multiplicity of each site is fixed between endmembers. This allows the implementation of Temkin-type models (Temkin, 1945), which are used to model melts. For example, here is the ideal part of the melt model proposed by Holland et al. (2018), restricted to the CMAS system (nonideality and changes in endmember energies have been ignored):"

This implies (to me) that an example block of code should following illustrating this type of model, but the next cell is another markdown cell with:

"The (a)symmetric model

Another solution model formalism implemented in BurnMan is the asymmetric model (Holland and Powell, 2003). This model is an extension of the regular solution model (Wohl, 1946), with the extension following van Laar (1906). ..."

Is the example missing, or am I missing something? If the latter, it might be worth clarifying.

In reference to: [openjournals/joss-reviews/issues/5389](https://github.com/openjournals/joss-reviews/issues/5389)
