### 2023-06-06 geodynamics/burnman#534 @simonwmatthews
tutorial_03: Suggestion for extra explanation in planet Zog creation
In tutorial 3, when creating the planet zog, an output file from perplex is imported to describe the convecting mantle. I think it would be useful to add some explanation of why importing the perplex output is advantageous over defining the convecting mantle layer in terms of burnman mineral objects.

This is in reference to [openjournals/joss-reviews/issues/5389](https://github.com/openjournals/joss-reviews/issues/5389)
