### 2023-06-15 f5d8310
update for numba, numpy, scipy
