+++ burnman/eos/reciprocal_kprime.py
@@ -23,14 +23,14 @@ except ImportError:
         return fn
 
 
-@jit
+@jit(nopython=True)
 def _delta_PoverK_from_P(PoverK, pressure, K_0, Kprime_0, Kprime_inf):
     return PoverK - (pressure / K_0) * np.power(
         (1.0 - Kprime_inf * PoverK), Kprime_0 / Kprime_inf
     )  # eq. 58
 
 
-@jit
+@jit(nopython=True)
 def _delta_PoverK_from_V(PoverK, V, V_0, K_0, Kprime_0, Kprime_inf):
     Kprime_ratio = Kprime_0 / Kprime_inf
     return (
