### 2023-06-16 8aa1402
relax example_fit_solution output
