+++ examples/example_fit_solution.py
@@ -191,7 +191,9 @@ if __name__ == "__main__":
     print("\nParameters:")
     print(fitted_eos.popt)
     print("\nFull covariance matrix:")
+    np.set_printoptions(precision=0)
     print(fitted_eos.pcov)
+    np.set_printoptions(precision=1)
     print("\nGoodness of fit:")
     print(fitted_eos.goodness_of_fit)
     print("\n")
@@ -238,7 +240,9 @@ if __name__ == "__main__":
     print("\nParameters:")
     print(fitted_eos.popt)
     print("\nFull covariance matrix:")
+    np.set_printoptions(precision=0)
     print(fitted_eos.pcov)
+    np.set_printoptions(precision=1)
     print("\nGoodness of fit:")
     print(fitted_eos.goodness_of_fit)
     print("\n")
