+++ burnman/__init__.py
@@ -81,7 +81,7 @@ Structure
 Requirements
 ------------
 
-  - Python 3.7+
+  - Python 3.8+
   - Python modules: NumPy, SciPy, SymPy, Matplotlib
   - Optional modules: cvxpy, pycddlib
 
