+++ setup.py
@@ -31,10 +31,10 @@ metadata = dict(
     package_data={"burnman": ["data/input_*/*"]},
     classifiers=[
         "License :: OSI Approved :: GNU General Public " "License v2 or later (GPLv2+)",
-        "Programming Language :: Python :: 3.7",
         "Programming Language :: Python :: 3.8",
         "Programming Language :: Python :: 3.9",
         "Programming Language :: Python :: 3.10",
+        "Programming Language :: Python :: 3.11",
     ],
 )
 
