+++ .github/workflows/main.yml
@@ -10,7 +10,7 @@ jobs:
 
     strategy:
       matrix:
-        python-versions: ['3.10']
+        python-versions: ['3.11']
 
     steps:
     - uses: actions/checkout@v2
@@ -38,7 +38,7 @@ jobs:
 
     strategy:
       matrix:
-        python-versions: ['3.7', '3.8', '3.9', '3.10']
+        python-versions: ['3.8', '3.9', '3.10', '3.11']
 
     steps:
     - uses: actions/checkout@v2
