### 2023-06-16 078f5d1
added useful PerpleX messages
