### 2023-06-16 1b05d04
improved composition fitting documentation, added MELTS ref
