+++ paper/paper.bib
@@ -495,3 +495,18 @@ year={2016},
 doi={10.3847/0004-637X/819/1/32},
 publisher={IOP Publishing}
 }
+
+@INPROCEEDINGS{Ghiorso:2007,
+       author = {{Ghiorso}, M.~S. and {Hirschmann}, M.~M. and {Grove}, T.~L.},
+        title = "{xMELTS: A thermodynamic model for the estimation of magmatic phase relations over the pressure range 0-30 GPa and at temperatures up to 2500 C}",
+     keywords = {3611 Thermodynamics (0766, 1011, 8411), 3612 Reactions and phase equilibria (1012, 8412), 3618 Magma chamber processes (1036), 3621 Mantle processes (1038), 3630 Experimental mineralogy and petrology},
+    booktitle = {AGU Fall Meeting Abstracts},
+         year = 2007,
+       volume = {2007},
+        month = dec,
+          eid = {V31C-0608},
+        pages = {V31C-0608},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2007AGUFM.V31C0608G},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
