### 2023-06-16 geodynamics/burnman#526 @bobmyhill
Thanks @kaylai

The manuscript text has now been updated to:
- clarify that this is constrained least squares
- reference the function used
- explain how the residuals and covariances are calculated
- explain how iron loss can be assessed by adding another phase to the minimization

See paper artifact here: https://github.com/geodynamics/burnman/actions/runs/5292256319
