+++ paper/paper.md
@@ -96,7 +96,9 @@ seismological and geodynamic software, including: Mineos [@Masters:2011],
 AxiSEM [@NissenMeyer:2014] and
 ASPECT [@Kronbichler:2012;@aspect-doi-v2.4.0;@aspectmanual]
 
-The project also includes a tutorial, a large collection of annotated examples,
+The project also includes a multipart tutorial
+(\url{https://burnman.readthedocs.io/en/latest/tutorial.html}),
+a large collection of annotated examples,
 an extensive suite of unit tests and benchmarks, and 
 a directory containing user-contributed code from published papers.
 Use of the code requires only moderate Python skills, and its modular nature
@@ -125,18 +127,21 @@ mineral database [@Stixrude:2011] as a foundation. Since then,
 its scope has expanded considerably. `BurnMan` now contains equations
 of state for minerals and melts from several published datasets.
 A common set of methods for all equations of state allows easy
-access to many thermodynamic properties
-(e.g. \autoref{fig:qtzproperties}), including anisotropic properties
-[@Myhill:2022].
+access to many thermodynamic properties, including anisotropic properties
+[@Myhill:2022]. A simple example is shown in \autoref{fig:qtzproperties}.
+See \url{https://burnman.readthedocs.io/en/latest/tutorial_01_material_classes.html#}
+for more details and the code required to make that plot.
 
 ![Heat capacity and bulk sound velocities of quartz through the alpha-beta
 quartz transition as found in [@Stixrude:2011]. This transition is modelled
-via a Landau-type model.
+via a Landau-type model. 
 \label{fig:qtzproperties}](figures/quartz_properties.png)
 
 Several other object classes in `BurnMan` build on the endmember
 equations of state. Several solution model formulations are included
-(with one example model shown in \autoref{fig:garnetsolution}),
+(with one example model shown in \autoref{fig:garnetsolution}, see
+\url{https://burnman.readthedocs.io/en/latest/tutorial_01_material_classes.html#}
+for more details),
 including both Gibbs and Helmholtz formulations [@Myhill:2018].
 Functions are also provided to convert solution models from one
 endmember basis to another [@Myhill:2021]. A composite material model
@@ -150,21 +155,21 @@ Gibbs minimization, while the endmember activities can be used to
 determine equilibrium via the equilibrium relations [@Holland:1998].
 \label{fig:garnetsolution}](figures/mg_ca_gt_properties.png)
 
-`BurnMan` includes many higher level functions that use
-its material model classes. These include functions to
-fit thermodynamic models for endmembers
-and solutions including full error propagation (\autoref{fig:fit}),
-construction of self-consistent planetary models (\autoref{fig:zog})
-and the calculation of seismic properties. Tools are provided to
-compare calculated seismic properties with published seismic models of
+Another class implemented in `BurnMan` is the `Composition` class
+which provides a framework to convert between mass, molar,
+and elemental compositions, convert to different component systems,
+and add or subtract chemical components (see
+\url{https://burnman.readthedocs.io/en/latest/tutorial_02_composition_class.html#}). 
+
+`BurnMan` also includes planetary `Layer` and `Planet` classes,
+which can be used to construct planetary models with self-consistent
+pressure, gravity and density profiles
+and calculate seismic properties through those bodies 
+(\autoref{fig:zog}, \url{https://burnman.readthedocs.io/en/latest/tutorial_03_layers_and_planets.html}).
+Tools are provided to compare those seismic properties with published seismic models of
 the Earth, and to produce input files to compute synthetic
 seismic data using other codes, including
[…]
