### 2023-06-16 geodynamics/burnman#527 @bobmyhill
I've now added references (and a link) to each of the tutorials at appropriate places in the manuscript text.
