### 2023-06-16 96e9218
add detail on optional dependencies to Readme
