+++ burnman/optimize/nonlinear_fitting.py
@@ -29,74 +29,53 @@ def nonlinear_least_squares_fit(
 
     Parameters
     ----------
-    model : class instance
-        Must have the following attributes:
-            data : 2D numpy array.
-                Elements of x[i][j] contain the observed position of
-                data point i
+    :param model: Model containing data to be fit, and functions to
+        aid in fitting.
+    :type model: object
 
-            data_covariances : 3D numpy array
-                Elements of cov[i][j][k] contain the covariance matrix
-                of data point i
+    :param lm_damping: Levenberg-Marquardt parameter for least squares minimization.
+    :type lm_damping: float
 
-            mle_tolerances : numpy array
-                The iterations to find the maximum likelihood estimator
-                for each observed data point will stop when mle_tolerances[i] <
-                np.linalg.norm(data_mle[i] - model.function(data_mle[i], flag))
-
-            delta_params : numpy array
-                parameter perturbations used to compute the jacobian
-
-        Must also contain the following functions:
-            set_params(self, param_values):
-                Function to set parameters
-
-            get_params(self):
-                Function to get current model parameters
-
-            function(self, x):
-                Returns value of model function evaluated at x
-
-            normal(self, x):
-                Returns value of normal to the model function
-                evaluated at x
-
-    lm_damping : float (optional, default: 0)
-        Levenberg-Marquardt parameter for least squares minimization
-
-    param_tolerance : float (optional, default: 1.e-5)
-        Levenberg-Marquardt iterations are terminated when
+    :param param_tolerance: Levenberg-Marquardt iterations are terminated when
         the maximum fractional change in any of the parameters
         during an iteration drops below this value
-
-    max_lm_iterations : integer (optional, default: 100)
-        Maximum number of Levenberg-Marquardt iterations
-
-    verbose : bool
-        Print some information to standard output
+    :type param_tolerance: float
 
 
-    Attributes added to model
-    ----------
-    n_dof : integer
-        Degrees of freedom of the system
-    data_mle : 2D numpy array
-        Maximum likelihood estimates of the observed data points
-        on the best-fit curve
-    jacobian : 2D numpy array
-        d(weighted_residuals)/d(parameter)
-    weighted_residuals : numpy array
-        Weighted residuals
-    weights : numpy array
-        1/(data variances normal to the best fit curve)
-    WSS : float
-        Weighted sum of squares residuals
-    popt : numpy array
-        Optimized parameters
-    pcov : 2D numpy array
-        Covariance matrix of optimized parameters
-    noise_variance : float
-        Estimate of the variance of the data normal to the curve
+    :param max_lm_iterations: Maximum number of Levenberg-Marquardt iterations
+    :type max_lm_iterations: int
+
+    :param verbose: Print some information to standard output
+    :type verbose: bool
+
+    .. note:: The object passed as model must have the following attributes:
+        * data [2D numpy.array] - Elements of x[i][j] contain the
+            observed position of data point i.
+        * data_covariances [3D numpy.array] Elements of cov[i][j][k] contain
+            the covariance matrix of data point i.
+        * mle_tolerances [numpy.array] - The iterations to find the maximum likelihood
+            estimator for each observed data point will stop when mle_tolerances[i] <
+            np.linalg.norm(data_mle[i] - model.function(data_mle[i], flag))
+        * delta_params [numpy.array] - parameter perturbations used to compute the jacobian
+
+        Must also have the following methods:
+        * set_params(self, param_values) -  Function to set parameters.
+        * get_params(self) - Function to get current model parameters.
[…]
