+++ test.sh
@@ -149,6 +149,7 @@ do
     [ $test == "gen_doc.py" ] && echo "  *** skipping $test !" && continue
     [ $test == "table_mineral_library.py" ] && echo "  *** skipping $test !" && continue
     [ $test == "create_burnman_readable_perplex_table.py" ] && echo "  *** skipping $test !" && continue
+    [ $test == "create_burnman_readable_perplex_table_v6.py" ] && echo "  *** skipping $test !" && continue
 
     testit $test $fulldir || exit 1
 done

[…]
