### 2023-06-18 geodynamics/burnman#529 @bobmyhill
Thanks @kaylai, this is now addressed by #541.
