+++ burnman/version.py
@@ -1,2 +1,0 @@
-version = "1.2.0a0"
-short_version = "1.2.0"
