### 2023-06-18 532b5fb
modify github yaml and test.sh to account for removed requirements file
