### 2023-06-18 geodynamics/burnman#532 @bobmyhill
Thanks! I've now:
- Updated burnman, fixing the compatibility issues
- Specified compatible versions of all modules in `pyproject.toml`
- Enforced a specific version of the black formatter (required for the indent script)
