### 2023-06-18 fb4fe9d
added other outputs for readthedocs
