+++ .readthedocs.yaml
@@ -4,6 +4,8 @@
 # Required
 version: 2
 
+formats: all
+
 # Set the version of Python and other tools you might need
 build:
   os: ubuntu-22.04
