### 2023-06-18 0585933
fix path to logo for pdf cover
