### 2023-06-19 cb639b3
updated equilibrium documentation
