+++ burnman/tools/equilibration.py
@@ -178,7 +178,7 @@ def F(
     reduced_free_composition_vectors,
 ):
     """
-    Returns a vector of values which are zero at equilibrium.
+    The vector-valued function for which the root is sought.
     The first two vector values depend on the
     equality_constraints chosen. For example, if
       - eq[i][0] = 'P', F[i] = P - eq[i][1]
@@ -254,9 +254,10 @@ def F(
 
 def jacobian(x, assemblage, equality_constraints, reduced_free_composition_vectors):
     """
-    The Jacobian of the equilibrium problem (dF/dx).
-    See documentation for F and get_parameters
-    (which return F and x respectively) for more details.
+    The Jacobian of the vector-valued function :math:`F` for which the
+    root is sought (:math:`\\partial F / \\partial x`).
+    See documentation for :func:`F` and :func:`get_parameters`
+    (which return :math:`F` and :math:`x` respectively) for more details.
 
     :param x: Parameter values for the equilibrium problem to be solved.
     :type x: numpy.array
@@ -472,6 +473,17 @@ def phase_fraction_constraints(phase, assemblage, fractions, prm):
     Converts a phase fraction constraint into standard linear form
     that can be processed by the root finding problem.
 
+    We start with a single fraction or an array of fractions
+    for a particular phase (:math:`n_p / \\sum n = f`).
+    These are then converted into the "X" form of constraint
+    by multiplying by :math:`\\sum n` and moving all terms to
+    the LHS of the equation:
+
+    :math:`-f n_0 - f n_1 - \\ldots + (1-f) n_p - \\ldots = 0`
+
+    This form is less readable, but easier to use as a constraint
+    in a nonlinear solve.
+
     :param phase: The phase for which the fraction is to be constrained
     :type phase: :class:`burnman.Solution` or :class:`burnman.Mineral`
 
@@ -507,17 +519,35 @@ def phase_composition_constraints(phase, assemblage, constraints, prm):
     that can be processed by the root finding problem.
 
     We start with constraints in the form (site_names, n, d, v), where
-    n*x/d*x = v and n and d are fixed vectors of site coefficients.
-    So, one could for example choose a constraint
+    :math:`(n x)/ (d x) = v` and :math:`n` and :math:`d` are fixed vectors of
+    site coefficients. So, one could for example choose a constraint
     ([Mg_A, Fe_A], [1., 0.], [1., 1.], [0.5]) which would
     correspond to equal amounts Mg and Fe on the A site.
 
-    These are then converted by this function into endmember proportions
-    (n'*p/d'*p = v). Because the proportions must add up to zero,
-    we can reexpress this ratio as a linear constraint:
-    [(n'[1:] - n'[0]) - v*(d'[1:] - d'[0])]*xi = v*d0 - n0
-    which is less easy for a human to understand
-    (in terms of chemical constraints), but easier to use as a constraint
+    This function converts the user-defined vectors of site constraints
+    :math:`n` and :math:`d` into vectors of endmember proportion
+    constraints :math:`n'` and :math:`d'`, such that
+    :math:`(n' x)/ (d' x) = v`. This is done via linear transformation
+    using the site occupancy matrix provided by :class:`burnman.Solution`.
+    By multiplying by the denominator, we have the following scalar
+    comparison: :math:`(n' x) = v (d' x)`
+
+    The equilibration function does not use the proportion of
+    the first endmember (as the endmember proportions must sum to one),
+    and so we split :math:`x`, :math:`n'` and :math:`d'` into the first
+    element and following elements:
+    :math:`(n'_0 x_0 + n'_i x_i) = v (d'_0 x_0 + d'_i x_i)`
+    where :math:`i` is taken over all elements apart from the first.
+
+    With some more rearranging we can express the constraint in standard
+    linear form:
[…]
