### 2023-06-19 geodynamics/burnman#530 @bobmyhill
To Do:
- [x] Update docstring
- [ ] Update manuscript
