### 2023-06-19 67857e1
improved equilibrium text
