+++ paper/paper.bib
@@ -510,3 +510,9 @@ publisher={IOP Publishing}
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@article{Nowak:1991,
+  title={A Family of Newton Codes for Systems of Highly Nonlinear Equations. Technical Report TR-91-10},
+  author={Nowak, U and Weimann, L},
+  journal={Zuse Institute Berlin, ZIB},
+  year={1991}
+}
