+++ docs/ref.bib
@@ -1271,3 +1271,10 @@
   adsurl   = {http://adsabs.harvard.edu/abs/1994JGR....99.4273A},
   adsnote  = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+@misc{Nowak1991,
+  title={A Family of Newton Codes for Systems of Highly Nonlinear Equations. Technical Report TR-91-10},
+  author={Nowak, U. and Weimann, L.},
+  journal={Zuse Institute Berlin, ZIB},
+  year={1991}
+}
