### 2023-06-19 geodynamics/burnman#530 @bobmyhill
Thanks! Both docstring and manuscript now updated. I've also added more functions to the autodocumentation so that the user can better trace all the calls.
