+++ paper/paper.bib
@@ -511,7 +511,7 @@ publisher={IOP Publishing}
 }
 
 @article{Nowak:1991,
-  title={A Family of Newton Codes for Systems of Highly Nonlinear Equations. Technical Report TR-91-10},
+  title="{A family of Newton codes for systems of highly nonlinear equations. Technical Report TR-91-10}",
   author={Nowak, U and Weimann, L},
   journal={Zuse Institute Berlin, ZIB},
   year={1991}
