+++ paper/paper.md
@@ -73,24 +73,25 @@ design that allows users to calculate many low-level properties that
 are not accessible using existing codes, and to combine various tools in
 novel, creative ways. The module includes:
 
-* many static and thermal equations of state for endmembers
-* several solution model formalisms (ideal, (a)symmetric, subregular)
+* over 10 static and thermal equations of state for endmembers
+* commonly-used solution model formalisms (ideal, (a)symmetric, subregular)
 and an option for user-defined functional forms
 * popular endmember and solution datasets 
 (including [@Holland:2011] and [@Stixrude:2011])
 * a consistent method for combining phases into a composite assemblage,
-with several seismic averaging schemes
+with seismic averaging schemes including Voigt, Reuss, Voigt-Reuss-Hill
+and the Hishin-Shtrikman bounds
 * a common set of methods to output thermodynamic and thermoelastic
 properties for all materials
 * an equilibrium reaction solver for composites
 * optimal least squares fitting routines for multivariate experimental
-data with (potentially correlated) errors. As an example, such functions
-can be used to simultaneously fit volumes, seismic velocities
-and enthalpies
-* a "Planet" class, which self-consistently calculates gravity profiles,
-mass, moment of inertia and temperature structure of planets given
+data with (potentially correlated) errors. Examples include simultaneous
+fitting of volumes, seismic velocities and enthalpies
+* a "Planet" class, which self-consistently calculates gravity-pressure-density
+profiles, mass, moment of inertia and temperature structure of planets given
 appropriate chemical and dynamic constraints
-* several geotherms from the literature
+* geothermal profiles from the literature as well as the option to calculate
+adiabatic profiles based on mineral assemblage
 * a set of high-level functions which create files readable by
 seismological and geodynamic software, including: Mineos [@Masters:2011],
 AxiSEM [@NissenMeyer:2014] and
@@ -98,7 +99,7 @@ ASPECT [@Kronbichler:2012;@aspect-doi-v2.4.0;@aspectmanual]
 
 The project also includes a multipart tutorial
 (\url{https://burnman.readthedocs.io/en/latest/tutorial.html}),
-a large collection of annotated examples,
+a over 40 annotated examples,
 an extensive suite of unit tests and benchmarks, and 
 a directory containing user-contributed code from published papers.
 Use of the code requires only moderate Python skills, and its modular nature
@@ -123,14 +124,20 @@ These are the needs satisfied by the `BurnMan` module.
 # The BurnMan project
 When `BurnMan` was first released [@Cottaar:2014], its focus was on
 the seismic properties of the lower mantle, using a single endmember
-mineral database [@Stixrude:2011] as a foundation. Since then,
-its scope has expanded considerably. `BurnMan` now contains equations
+mineral database [@Stixrude:2011] as a foundation.
+This initial release had no solution model implementation,
+no model for Gibbs energy or other thermodynamic potentials,
+no higher level functionality for fitting of experimental data or
+inversions, and no functions designed to model planetary bodies.
+Since then, its scope has expanded considerably
+(see \url{https://github.com/geodynamics/burnman/releases}
+for headline improvements). `BurnMan` now contains equations
 of state for minerals and melts from several published datasets.
 A common set of methods for all equations of state allows easy
 access to many thermodynamic properties, including anisotropic properties
 [@Myhill:2022]. A simple example is shown in \autoref{fig:qtzproperties}.
 See \url{https://burnman.readthedocs.io/en/latest/tutorial_01_material_classes.html#}
-for more details and the code required to make that plot.
+for more details and the code required to make the plot.
 
[…]
