### 2023-06-19 0b3c892
only include important eqm functions in docs
