+++ paper/paper.bib
@@ -496,19 +496,23 @@ doi={10.3847/0004-637X/819/1/32},
 publisher={IOP Publishing}
 }
 
-@INPROCEEDINGS{Ghiorso:2007,
-       author = {{Ghiorso}, M.~S. and {Hirschmann}, M.~M. and {Grove}, T.~L.},
-        title = "{xMELTS: A thermodynamic model for the estimation of magmatic phase relations over the pressure range 0-30 GPa and at temperatures up to 2500 C}",
-     keywords = {3611 Thermodynamics (0766, 1011, 8411), 3612 Reactions and phase equilibria (1012, 8412), 3618 Magma chamber processes (1036), 3621 Mantle processes (1038), 3630 Experimental mineralogy and petrology},
-    booktitle = {AGU Fall Meeting Abstracts},
-         year = 2007,
-       volume = {2007},
-        month = dec,
-          eid = {V31C-0608},
-        pages = {V31C-0608},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2007AGUFM.V31C0608G},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
+@article{Ghiorso:1995,
+	author = {Ghiorso, Mark S. and Sack, Richard O.},
+	date = {1995/03/01},
+	date-added = {2023-06-19 18:07:56 +0100},
+	date-modified = {2023-06-19 18:07:56 +0100},
+	doi = {10.1007/BF00307281},
+	id = {Ghiorso1995},
+	isbn = {1432-0967},
+	journal = {Contributions to Mineralogy and Petrology},
+	number = {2},
+	pages = {197--212},
+	title = "{Chemical mass transfer in magmatic processes IV. A revised and internally consistent thermodynamic model for the interpolation and extrapolation of liquid-solid equilibria in magmatic systems at elevated temperatures and pressures}",
+	url = {https://doi.org/10.1007/BF00307281},
+	volume = {119},
+	year = {1995},
+	bdsk-url-1 = {https://doi.org/10.1007/BF00307281}}
+
 
 @article{Nowak:1991,
   title="{A family of Newton codes for systems of highly nonlinear equations. Technical Report TR-91-10}",
