+++ burnman/minerals/SLB_2022.py
@@ -1,0 +1,1950 @@
+# This file is part of BurnMan - a thermoelastic and thermodynamic toolkit
+# for the Earth and Planetary Sciences
+# Copyright (C) 2012 - 2023 by the BurnMan team,
+# released under the GNU GPL v2 or later.
+
+
+"""
+SLB_2022
+Minerals from Stixrude & Lithgow-Bertelloni 2022 and references therein
+File autogenerated using HeFESTO_to_burnman.py and HeFESTo_parameters_010121
+"""
+
+from __future__ import absolute_import
+
+
+from ..classes.mineral import Mineral
+from ..classes.solution import Solution
+from ..classes.solutionmodel import (
+    IdealSolution,
+    SymmetricRegularSolution,
+    AsymmetricRegularSolution,
+)
+
+
+"""
+ENDMEMBERS
+"""
+
+
+class ab(Mineral):
+    def __init__(self):
+        self.params = {
+            "name": "Albite",
+            "formula": {"Na": 1.0, "Al": 1.0, "Si": 3.0, "O": 8.0},
+            "equation_of_state": "slb3",
+            "F_0": -3.7201476,
+            "V_0": 0.000100452,
+            "K_0": 59752680000.0,
+            "Kprime_0": 2.77841,
+            "Debye_0": 720.00359,
+            "grueneisen_0": 0.57885,
+            "q_0": 1.0,
+            "G_0": 36000000000.0,
+            "Gprime_0": 1.38571,
+            "eta_s_0": 1.02978,
+            "n": 13.0,
+            "Z": 4.0,
+            "molar_mass": 0.262223,
+        }
+
+        Mineral.__init__(self)
+
+
+class al(Mineral):
+    def __init__(self):
+        self.params = {
+            "name": "Almandine",
+            "formula": {"Fe": 3.0, "Al": 2.0, "Si": 3.0, "O": 12.0},
+            "equation_of_state": "slb3",
+            "F_0": -4.93302797,
+            "V_0": 0.00011543,
+            "K_0": 173896370000.0,
+            "Kprime_0": 4.91341,
+            "Debye_0": 741.38227,
+            "grueneisen_0": 1.06493,
+            "q_0": 1.42169,
+            "G_0": 96000000000.0,
+            "Gprime_0": 1.40927,
+            "eta_s_0": 2.09289,
+            "n": 20.0,
+            "Z": 4.0,
+            "molar_mass": 0.49776,
+        }
+
+        self.property_modifiers = [
+            ["landau_slb_2022", {"Tc_0": 7.5, "S_D": 40.14405, "V_D": 0.0}]
+        ]
+
+        Mineral.__init__(self)
+
+
+class alpv(Mineral):
+    def __init__(self):
+        self.params = {
+            "name": "Al-Perovskite",
+            "formula": {"Al": 2.0, "O": 3.0},
+            "equation_of_state": "slb3",
+            "F_0": -1.51772863,
+            "V_0": 2.4944e-05,
+            "K_0": 242400000000.0,
+            "Kprime_0": 4.1,
+            "Debye_0": 858.26509,
+            "grueneisen_0": 1.54222,
+            "q_0": 0.84088,
+            "G_0": 169199620000.0,
+            "Gprime_0": 1.55703,
+            "eta_s_0": 2.2751,
+            "n": 5.0,
+            "Z": 4.0,
+            "molar_mass": 0.101961,
+        }
+
+        Mineral.__init__(self)
+
+
+class an(Mineral):
+    def __init__(self):
+        self.params = {
+            "name": "Anorthite",
+            "formula": {"Ca": 1.0, "Al": 2.0, "Si": 2.0, "O": 8.0},
[…]
