+++ burnman/data/input_raw_endmember_datasets/HeFESTo_to_burnman.py
@@ -1,0 +1,424 @@
+# This file is part of BurnMan - a thermoelastic and thermodynamic toolkit for
+# the Earth and Planetary Sciences
+# Copyright (C) 2012 - 2023 by the BurnMan team, released under the GNU
+# GPL v2 or later.
+
+
+# This is a script that converts a version of the
+# HeFESTo data format into the standard burnman
+# format (printed to stdout)
+# This script is used to generate the 2022 dataset,
+# taking parameters from https://github.com/stixrude/HeFESTo_parameters_010121
+# via git clone https://github.com/stixrude/HeFESTo_parameters_010121.git
+# The SLB_2022 BurnMan file is then created via the commands
+# python HeFESTo_to_burnman.py > ../../minerals/SLB_2022.py; black ../../minerals/SLB_2022.py
+# and from . import SLB_2022 added to the minerals __init__.py
+
+
+import os
+import pathlib
+import pandas as pd
+import numpy as np
+from burnman.utils.chemistry import dictionarize_formula
+from itertools import groupby
+import re
+
+
+def all_equal(iterable):
+    g = groupby(iterable)
+    return next(g, True) and not next(g, False)
+
+
+def rfloat(x, m=1.0):
+    return round(float(x) * m, ndigits=10)
+
+
+hefesto_path = "HeFESTo_parameters_010121"
+mbrdir = pathlib.Path(hefesto_path)
+phasedir = pathlib.Path(f"{hefesto_path}/phase")
+
+param_dict = {}
+with open("hefesto_parameter_names.txt") as file:
+    lines = [line.rstrip() for line in file]
+    for line in lines:
+        param_dict[" ".join(line.split()[1:])] = line.split()[0]
+
+ignored_mbrs = [
+    "crst",
+    "enm",
+    "fapv",
+    "fea",
+    "fee",
+    "feg",
+    "flpv",
+    "hem",
+    "hepv",
+    "hlpv",
+    "hmag",
+    "hppv",
+    "lppv",
+    "mag",
+    "mgl",
+    "sil",
+    "wuls",
+]
+
+solution_aliases = {
+    "sp": "mg_fe_aluminous_spinel",
+    "pv": "bridgmanite",
+    "opx": "orthopyroxene",
+    "gt": "garnet",
+    "ol": "olivine",
+    "cf": "calcium_ferrite_structured_phase",
+    "ppv": "post_perovskite",
+    "c2c": "c2c_pyroxene",
+    "ri": "ringwoodite",
+    "plg": "plagioclase",
+    "mw": "ferropericlase",
+    "cpx": "clinopyroxene",
+    "nal": "new_aluminous_phase",
+    "il": "ilmenite",
+    "wa": "wadsleyite",
+}
+
+# solutions
+sol_params = {}
+for f in phasedir.iterdir():
+    # Do not process liquid!!
+    if os.stat(f).st_size > 0 and str(f).split("/")[-1] != "liq":
+        name = str(f).split("/")[-1]
+        d = pd.read_csv(f, delim_whitespace=True)
+
+        endmembers = list(d.columns)
+
+        n_mbrs = len(endmembers)
+
+        data = d.values.tolist()
+
+        n_data = len(data)
+        v0 = n_data - n_mbrs
+
+        # Preprocess interaction matrices
+        # to get rid of ignored endmembers
+        use_indices = np.array(
+            [i for i, mbr in enumerate(endmembers) if mbr not in ignored_mbrs]
+        )
+        use_endmembers = np.array(
+            [mbr for mbr in endmembers if mbr not in ignored_mbrs]
+        )
+        n_use_mbrs = len(use_indices)
+
+        energy_matrix = np.array(
+            np.array(data)[np.ix_(use_indices, use_indices)], dtype=float
[…]

[…]
