+++ docs/changelog/20230620_bobmyhill.rst
@@ -1,0 +1,3 @@
+* BurnMan now has an implementation of the SLB2022 dataset.
+
+  *Bob Myhill, 2023/06/20*
