### 2023-06-20 b0eeb3a
fix units for the Helmholtz energy
