+++ burnman/minerals/SLB_2022.py
@@ -33,7 +33,7 @@ class ab(Mineral):
             "name": "Albite",
             "formula": {"Na": 1.0, "Al": 1.0, "Si": 3.0, "O": 8.0},
             "equation_of_state": "slb3",
-            "F_0": -3.7201476,
+            "F_0": -3720147.5999999996,
             "V_0": 0.000100452,
             "K_0": 59752680000.0,
             "Kprime_0": 2.77841,
@@ -57,7 +57,7 @@ class al(Mineral):
             "name": "Almandine",
             "formula": {"Fe": 3.0, "Al": 2.0, "Si": 3.0, "O": 12.0},
             "equation_of_state": "slb3",
-            "F_0": -4.93302797,
+            "F_0": -4933027.97,
             "V_0": 0.00011543,
             "K_0": 173896370000.0,
             "Kprime_0": 4.91341,
@@ -85,7 +85,7 @@ class alpv(Mineral):
             "name": "Al-Perovskite",
             "formula": {"Al": 2.0, "O": 3.0},
             "equation_of_state": "slb3",
-            "F_0": -1.51772863,
+            "F_0": -1517728.6300000001,
             "V_0": 2.4944e-05,
             "K_0": 242400000000.0,
             "Kprime_0": 4.1,
@@ -109,7 +109,7 @@ class an(Mineral):
             "name": "Anorthite",
             "formula": {"Ca": 1.0, "Al": 2.0, "Si": 2.0, "O": 8.0},
             "equation_of_state": "slb3",
-            "F_0": -4.01336904,
+            "F_0": -4013369.04,
             "V_0": 0.00010061,
             "K_0": 84093310000.0,
             "Kprime_0": 6.73404,
@@ -133,7 +133,7 @@ class anao(Mineral):
             "name": "alpha-NaO2_phase",
             "formula": {"Na": 2.0, "Al": 2.0, "O": 4.0},
             "equation_of_state": "slb3",
-            "F_0": -2.11483579,
+            "F_0": -2114835.79,
             "V_0": 4.542e-05,
             "K_0": 161143930000.0,
             "Kprime_0": 3.90838,
@@ -157,7 +157,7 @@ class apbo(Mineral):
             "name": "alpha-PbO_2-SiO_2",
             "formula": {"Si": 1.0, "O": 2.0},
             "equation_of_state": "slb3",
-            "F_0": -0.79336684,
+            "F_0": -793366.84,
             "V_0": 1.367e-05,
             "K_0": 327156010000.0,
             "Kprime_0": 4.01662,
@@ -181,7 +181,7 @@ class appv(Mineral):
             "name": "Al-Post-Perovskite",
             "formula": {"Al": 2.0, "O": 3.0},
             "equation_of_state": "slb3",
-            "F_0": -1.33646473,
+            "F_0": -1336464.73,
             "V_0": 2.3847e-05,
             "K_0": 249000000000.0,
             "Kprime_0": 4.0,
@@ -205,7 +205,7 @@ class capv(Mineral):
             "name": "Ca-Perovskite",
             "formula": {"Ca": 1.0, "Si": 1.0, "O": 3.0},
             "equation_of_state": "slb3",
-            "F_0": -1.45991018,
+            "F_0": -1459910.1800000002,
             "V_0": 2.745e-05,
             "K_0": 236000000000.0,
             "Kprime_0": 3.9,
@@ -229,7 +229,7 @@ class cats(Mineral):
             "name": "Lime_Tschermak'",
             "formula": {"Ca": 1.0, "Al": 2.0, "Si": 1.0, "O": 6.0},
             "equation_of_state": "slb3",
[…]
