### 2023-06-21 #5389 @kaylai
Agreed, this journal review experience has been excellent, and I will recommend JOSS to my colleagues!
