### 2023-06-21 #5389 @jedbrown
Thanks for the excellent and timely reviews!

@bobmyhill I have a couple editorial comments on the paper:
1. This is a trivial fix and removing the narrative reference to the repository since it's already in the sidebar.
```diff
diff --git c/paper/paper.md w/paper/paper.md
index 9d16405..930ec9f 100644
--- c/paper/paper.md
+++ w/paper/paper.md
@@ -77,7 +77,7 @@ novel, creative ways. The module includes:
 * commonly-used solution model formalisms (ideal, (a)symmetric, subregular)
 and an option for user-defined functional forms
 * popular endmember and solution datasets
-(including [@Holland:2011] and [@Stixrude:2011])
+(including @Holland:2011 and @Stixrude:2011)
 * a consistent method for combining phases into a composite assemblage,
 with seismic averaging schemes including Voigt, Reuss, Voigt-Reuss-Hill
 and the Hishin-Shtrikman bounds
@@ -104,8 +104,6 @@ an extensive suite of unit tests and benchmarks, and
 a directory containing user-contributed code from published papers.
 Use of the code requires only moderate Python skills, and its modular nature
 means that it can easily be customised.
-The `BurnMan` project can be found at
-\url{https://github.com/geodynamics/burnman}.

 # Statement of need
 Earth Scientists are interested in a number of different material properties,
```
2. The paper is "long" by JOSS standards and I think parts are duplicative of the project documentation, which is an explicit non-goal of JOSS. The project history bit isn't usually included in JOSS papers and the heavy URL linking into project/API docs feels especially duplicative. This is also a concern because those links are potentially unstable on archival scales. Would you be willing to tighten this section just a bit into highlighting key capabilities more than being a link aggregator into the project docs?
3. Are you familiar with [CoolProp](https://github.com/CoolProp/CoolProp)? In my reading through, it seems like there's significant overlap in methods and positioning, though it focuses on different materials. In your opinion, is it close enough to be worth citing?
