+++ paper/paper.md
@@ -55,105 +55,97 @@ bibliography: paper.bib
 Many branches of physics, chemistry and Earth sciences build
 complex materials from simpler constructs: rocks are composites
 made up of one or more phases; phases are solutions of more than one
-endmember and endmembers are usually mixtures of more than one element.
+endmember; and endmembers are usually mixtures of more than one element.
 The properties of the endmember building blocks at different pressures and
-temperatures can be provided by a wide array of different equations of state.
-The averaging of the endmember properties within solutions and composite
-materials can also be achieved in several different ways. Once calculated,
+temperatures rovided modelled using a wide array of different equations of state.
+There are also many models for the averaging of endmember properties
+within solutions and composite materials. Once calculated,
 the physical properties of composite materials can be used in many
 different ways.
 
-
-`BurnMan` is an open source, extensible mineral physics Python module
-written in Python to provide a well-tested, benchmarked toolbox that
-is able to use several different methods to calculate the
-physical properties of natural materials.
+`BurnMan` is an open source, extensible mineral physics module
+written in Python. It implements several different methods to
+calculate the physical properties of natural materials.
 The toolbox has a class-based, modular
 design that allows users to calculate many low-level properties that
-are not accessible using existing codes, and to combine various tools in
+are not easily accessed using existing codes, and to combine various tools in
 novel, creative ways. The module includes:
 
-* over 10 static and thermal equations of state for endmembers
+* over a dozen static and thermal equations of state for pure phases;
 * commonly-used solution model formalisms (ideal, (a)symmetric, subregular)
-and an option for user-defined functional forms
-* popular endmember and solution datasets 
-(including [@Holland:2011] and [@Stixrude:2011])
+and a formalism that allows users to define their own excess energy functions;
+* popular endmember and solution datasets for solids and melts,
+including @Holland:2011, @deKoker:2013 and @Stixrude:2022;
+* an anisotropic equation of state [@Myhill:2022];
 * a consistent method for combining phases into a composite assemblage,
 with seismic averaging schemes including Voigt, Reuss, Voigt-Reuss-Hill
-and the Hishin-Shtrikman bounds
+and the Hishin-Shtrikman bounds;
 * a common set of methods to output thermodynamic and thermoelastic
-properties for all materials
-* an equilibrium reaction solver for composites
+properties for all materials;
+* an solver to chemically equilibrate composite materials;
 * optimal least squares fitting routines for multivariate experimental
-data with (potentially correlated) errors. Examples include simultaneous
-fitting of volumes, seismic velocities and enthalpies
-* a "Planet" class, which self-consistently calculates gravity-pressure-density
-profiles, mass, moment of inertia and temperature structure of planets given
-appropriate chemical and dynamic constraints
+data with (potentially correlated) errors. These allow (for example) simultaneous
+fitting of pure phase and solution model parameters to experimental volumes,
+seismic velocities and enthalpies of formation;
+* "Planet" and "Layer" classes that self-consistently calculate
+gravity, pressure, density, mass, moment of inertia and seismic velocity profiles
+given chemical, thermal and dynamic constraints;
 * geothermal profiles from the literature as well as the option to calculate
-adiabatic profiles based on mineral assemblage
+adiabatic profiles based on mineral assemblage;
 * a set of high-level functions which create files readable by
 seismological and geodynamic software, including: Mineos [@Masters:2011],
 AxiSEM [@NissenMeyer:2014] and
[…]
