+++ paper/paper.bib
@@ -99,6 +99,18 @@ url = {https://doi.org/10.2138/am.2010.3354},
 eprint = {https://pubs.geoscienceworld.org/msa/ammin/article-pdf/95/7/1006/3626686/14\_3354DeCapitani.pdf},
 }
 
+@article{deKoker:2013,
+  author  = {{de Koker}, N. and {Karki}, B.~B. and {Stixrude}, L.},
+  title   = {{Thermodynamics of the MgO-SiO$_{2}$ liquid system in Earth's lowermost mantle from first principles}},
+  journal = {Earth and Planetary Science Letters},
+  year    = 2013,
+  month   = jan,
+  volume  = 361,
+  pages   = {58-63},
+  doi     = {10.1016/j.epsl.2012.11.026},
+  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
 @article{Stixrude:2022,
 author = {Stixrude, Lars and Lithgow-Bertelloni, Carolina},
 title = "{Thermal expansivity, heat capacity and bulk modulus of the mantle}",
@@ -519,4 +531,21 @@ publisher={IOP Publishing}
   author={Nowak, U and Weimann, L},
   journal={Zuse Institute Berlin, ZIB},
   year={1991}
+}
+
+@article{Diamond:2016,
+       author = {{Diamond}, Steven and {Boyd}, Stephen},
+        title = "{CVXPY: A Python-Embedded Modeling Language for Convex Optimization}",
+      journal = {arXiv e-prints},
+     keywords = {Mathematics - Optimization and Control},
+         year = 2016,
+        month = mar,
+          eid = {arXiv:1603.00943},
+        pages = {arXiv:1603.00943},
+          doi = {10.48550/arXiv.1603.00943},
+archivePrefix = {arXiv},
+       eprint = {1603.00943},
+ primaryClass = {math.OC},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2016arXiv160300943D},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
