### 2023-06-22 #5389 @bobmyhill
Thanks @jedbrown, I've implemented your suggestions from (1) and (2). I also cited cvxpy and a couple of other important dataset studies that I originally missed.

I had never heard of coolprop; thank you for bringing it to my attention! I agree that it shares some goals with BurnMan. I thought for a while about where I could cite it in this paper, but I just can't find a place where it would go. It is focused on low pressure fluids for chemical applications. So I don't think it is close enough to cite for this paper. If in the future we expand our scope to crustal fluids, then I think it will be worth revisiting this.

I'm finished making changes now, but happy to address any more comments you might have. Thanks again!
