+++ burnman/__init__.py
@@ -172,11 +172,16 @@ Citing BurnMan
 
 If you use BurnMan in your work, we ask that you cite the following publications:
 
-  - Myhill, R., Cottaar, S., Heister, T., Rose, I., and Unterborn, C. (2022):
-    BurnMan v1.1.0 [Software]. Computational Infrastructure for Geodynamics. Zenodo.
-    `(https://doi.org/10.5281/zenodo.7080174) <https://doi.org/10.5281/zenodo.7080174>`_
+  - Myhill, R., Cottaar, S., Heister, T., Rose, I., Unterborn, C.,
+    Dannberg, J. and Gassmoeller, R. (2023). BurnMan - a Python toolkit for
+    planetary geophysics, geochemistry and thermodynamics. Journal of Open Source Software.
+    `(https://doi.org/10.21105/joss.05389) <https://doi.org/10.21105/joss.05389>`_
 
-  - Cottaar S., Heister, T., Rose, I., and Unterborn, C., 2014, BurnMan: A
+  - Myhill, R., Cottaar, S., Heister, T., Rose, I., and Unterborn, C. (2023):
+    BurnMan v1.2.0 [Software]. Computational Infrastructure for Geodynamics. Zenodo.
+    `(https://doi.org/10.5281/zenodo.8104293) <https://doi.org/10.5281/zenodo.8104293>`_
+
+  - Cottaar S., Heister, T., Rose, I., and Unterborn, C., (2014). BurnMan: A
     lower mantle mineral physics toolkit, Geochemistry, Geophysics, and
     Geosystems, 15(4), 1164-1179 `(https://doi.org/10.1002/2013GC005122)
     <https://doi.org/10.1002/2013GC005122>`_
