### 2023-07-01 #5389 @bobmyhill
@jedbrown  Thanks! I hope this is what you need:

Tagged version: [v1.2.0](https://github.com/geodynamics/burnman/releases/tag/v1.2)
Zenodo archive: https://doi.org/10.5281/zenodo.8104293
DOI: 10.5281/zenodo.8104293
