### 2023-07-01 #5389 @jedbrown
@bobmyhill I think you can just remove that `isbn` from `Ghiorso:1995` -- it has a DOI already and that other number is not an ISBN.
