### 2023-07-01 0d8d00b
removed troublesome metadata from Ghiorso reference
