+++ paper/paper.bib
@@ -510,20 +510,14 @@ publisher={IOP Publishing}
 
 @article{Ghiorso:1995,
 	author = {Ghiorso, Mark S. and Sack, Richard O.},
-	date = {1995/03/01},
-	date-added = {2023-06-19 18:07:56 +0100},
-	date-modified = {2023-06-19 18:07:56 +0100},
 	doi = {10.1007/BF00307281},
-	id = {Ghiorso1995},
-	isbn = {1432-0967},
 	journal = {Contributions to Mineralogy and Petrology},
 	number = {2},
 	pages = {197--212},
 	title = "{Chemical mass transfer in magmatic processes IV. A revised and internally consistent thermodynamic model for the interpolation and extrapolation of liquid-solid equilibria in magmatic systems at elevated temperatures and pressures}",
-	url = {https://doi.org/10.1007/BF00307281},
 	volume = {119},
-	year = {1995},
-	bdsk-url-1 = {https://doi.org/10.1007/BF00307281}}
+	year = {1995}
+  }
 
 
 @article{Nowak:1991,
