### 2023-07-01 #5389 @bobmyhill
@jedbrown Brilliant, the proof looks good to me, with the exception of a couple of typos (now fixed in the paper branch).
